--- crates/hammerslide-rs/src/lib.rs
@@ -1,3 +1,0 @@
-mod aggregation;
-mod circular_queue;
-mod hammer_slide;
